vm runs with cursor

+++ src/vm/debug_inst.c
@@ -6,7 +6,7 @@
 /*   By: mnishimo <mnishimo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/09 21:12:15 by mnishimo          #+#    #+#             */
-/*   Updated: 2019/06/16 23:33:55 by mnishimo         ###   ########.fr       */
+/*   Updated: 2019/06/17 14:22:22 by mnishimo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,10 +50,10 @@ void	deb_64_log(t_game *game, t_inst *inst, t_process *caller, int p_id)
 				p_id, inst->args[0].value.u_dir_val);
 	else if (opcode == e_zjmp && caller->carry)
 		l = ft_asprintf(&tmp, "zjmp to memdump(%i)!\n",
-				p_id, inst->args[0].value.u_dir_val);
+				p_id, (int)inst->args[0].value.u_dir_val);
 	else if (opcode == e_zjmp)
 		l = ft_asprintf(&tmp, "zjmp to memdump(%i) skipped!\n",
-				p_id, inst->args[0].value.u_dir_val);
+				p_id, (int)inst->args[0].value.u_dir_val);
 	else
 		return ;
 	add_head(&tmp, caller->p_id, caller->c_id, &l);
